Fix warnings and C++ error ...

+++ demo/Arrangement_on_surface_2/ArrangementGraphicsItem.h
@@ -371,11 +371,11 @@ paint(QPainter* painter, TTraits /* traits */)
 
     // Filling Arrangement info widget
     QArrangementInfoWidget* infoWidget = QArrangementInfoWidget::instance();
-    infoWidget->setNumFaces             ( this->arr->number_of_faces(), this->arr->number_of_unbounded_faces() );
-    infoWidget->setNumEdges             ( this->arr->number_of_edges() );
-    infoWidget->setNumVertices          ( this->arr->number_of_vertices() );
-    infoWidget->setNumIsolatedVertices  ( this->arr->number_of_isolated_vertices() );
-    infoWidget->setNumVerticesAtInfinity( this->arr->number_of_vertices_at_infinity() );
+    infoWidget->setNumFaces             ( (int)this->arr->number_of_faces(), (int)this->arr->number_of_unbounded_faces() );
+    infoWidget->setNumEdges             ( (int)this->arr->number_of_edges() );
+    infoWidget->setNumVertices          ( (int)this->arr->number_of_vertices() );
+    infoWidget->setNumIsolatedVertices  ( (int)this->arr->number_of_isolated_vertices() );
+    infoWidget->setNumVerticesAtInfinity( (int)this->arr->number_of_vertices_at_infinity() );
 
     this->painterostream = ArrangementPainterOstream< Traits >( painter, this->boundingRect( ) );
     this->painterostream.setScene( this->scene );
